% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_seurat.R
\name{AddSignatureScore}
\alias{AddSignatureScore}
\title{Calculates a signature score for a given set of features and adds it to the
object metadata.}
\usage{
AddSignatureScore(object, features, assay, slot, name)
}
\arguments{
\item{name}{}
}
\value{

}
\description{
Calculates a signature score for a given set of features and adds it to the
object metadata.
}
