% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCluster.R
\name{annofactor}
\alias{annofactor}
\title{Extract level of reprogramming cells.}
\usage{
annofactor(mat)
}
\arguments{
\item{mat}{Expression matrix.}
}
\value{
factor represents timepoints
}
\description{
Extract level of reprogramming cells.
}
