% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{apclusK}
\alias{apclusK}
\title{Clustering genes to k clusters using affinity propagation algorithm.}
\usage{
apclusK(mat, k = 7, metric = "pearson", negDist = NULL,
  cell_labels = NULL, check_outlier_pattern = FALSE, check_min_size = 5,
  check_cor_thr = -0.5, ...)
}
\arguments{
\item{mat}{Gene expression matrix, columns are cells and rows are genes.}

\item{k}{Desired number of gene clusters.}

\item{metric}{Metric used to compute the paired distance - default is pearson correlation distance.}

\item{negDist}{The dissimilarity matrix given by user.}

\item{cell_labels}{cell clustering labels, it must be given when check_outlier_pattern = TRUE.}

\item{check_outlier_pattern}{Detect and remove negtive expression pattern dominated by rare cells.}

\item{check_min_size}{Check the correlation of the cell cluster smaller than check_min_size with all gene clusters.}

\item{check_cor_thr}{Correlation threshold to remove outlier pattern.}

\item{...}{Additional arguments passed on to apclusterK.}
}
\value{
Gene clustering results.
}
\description{
Runs affinity propagation clustering.
}
