% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCluster.R
\name{aspecc}
\alias{aspecc}
\title{Spectral clustering.}
\usage{
aspecc(Dist, numk = NULL, ka = 10, kNN = NULL, type = 3,
  estimateRage = 3:8, scaleDist = TRUE)
}
\arguments{
\item{Dist}{Paire-wised distance matrix.}

\item{numk}{Number of clusters number. If NULL, estimate number will be used.}

\item{ka}{To equalize the number of neighbors we set the value sigma(i) for each cell i to the distance to its kath nearest neighbor. Default is 10.}

\item{kNN}{Additional affinities beyond Kth nearest neighbor are set to zero. If NULL, kNN will be set the 3ka.}

\item{type}{The variants of spectral clustering to use, see \code{\link[SNFtool]{spectralClustering}}.}

\item{estimateRage}{Estimate range.}

\item{scaleDist}{scale distance.}
}
\value{
Spectral clustering labels, and affinity matrix.
}
\description{
Clustering cells based on adaptive gauss kernel.
}
\references{
David van Dijk et al. MAGIC: A diffusion-based imputation method reveals gene-gene interactions in single-cell RNA-sequencing data. BioRxiv 2017.

Bo Wang et al. Similarity network fusion for aggregating data types on a genomic scale. Nature Methods 2014.
}
