% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bclone_processing.R
\name{Bclone_processing}
\alias{Bclone_processing}
\title{Process B cell clonality}
\usage{
Bclone_processing(ser)
}
\arguments{
\item{ser}{Seurat object containing T cell information}
}
\value{
Feature plots highlighting clones
}
\description{
This script reads in a counts file from either the DropSeq or 10x 
pipeline, converts the genes to a given naming convention (MGI or HGNC) and return a Seurat object
}
