% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bindByQuantiles}
\alias{bindByQuantiles}
\title{bindByQuantiles}
\usage{
bindByQuantiles(vec, q_low = 0, q_high = 0.99)
}
\arguments{
\item{vec}{}

\item{q_low}{}

\item{q_high}{}
}
\value{
...
}
\description{
bindByQuantiles
}
