% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_seurat.R
\name{BuildSeuratObject}
\alias{BuildSeuratObject}
\title{Builds Seurat object from raw data. The samples and data paths to the raw
Cellranger data are read from an excel manifest file.}
\usage{
BuildSeuratObject(file, sheet)
}
\arguments{
\item{sheet}{}
}
\value{

}
\description{
Builds Seurat object from raw data. The samples and data paths to the raw
Cellranger data are read from an excel manifest file.
}
