% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcKNNgraph}
\alias{calcKNNgraph}
\title{Calculate k Nearest Neighbors from some distance metric}
\usage{
calcKNNgraph(x, k)
}
\arguments{
\item{x}{An m x n numeric matrix}

\item{k}{The number of nearest neighbors to return
per sample}
}
\value{
An m x k matrix of indicies 1...k of the
nearest neighbors for the specified row.
}
\description{
This function returns a logical vector identifying if
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}

