% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{call.back}
\alias{call.back}
\title{Retrieve genes with high correlation with existing pattern.}
\usage{
call.back(mat, apres, q = 0.01, method = "BH", ...)
}
\arguments{
\item{mat}{Gene expression matrix, columns are cells and rows are genes.}

\item{apres}{Dataframe of affinity propagation clustering results.}

\item{q}{The treshold of adjust p-value of correlation coefficient. Genes with p-value less than q will be retrieved to the gene expression matrix}

\item{method}{Correction method for p-value, see \code{\link[stats]{p.adjust}}.}

\item{...}{Additional arguments passed on to \code{\link{p.cor}}.}
}
\value{
Dataframe with selected genes and found genes.
}
\description{
Retrieve genes with high correlation with the exemplars of affinity propagation clustering results.
}
