% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{ClusterCentroids}
\alias{ClusterCentroids}
\title{Cluster Centrois}
\usage{
ClusterCentroids(
  object,
  reduction.use = "pca",
  pcs.use = 1:10,
  cells.use = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.use}{dimensionality reduction to use}

\item{pcs.use}{Number of PCs to use}

\item{cells.use}{Group of cells to calculate centroids; default = all}
}
\description{
Identify cluster centroids in PC space; used by other functions
}
\examples{
ClusterCentroids()
}
\keyword{centroids}
