% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{ComputeClusterDistances}
\alias{ComputeClusterDistances}
\title{This function calculates next closest cluster for each cell; used for merging}
\usage{
ComputeClusterDistances(
  object,
  reduction.use = "pca",
  dist.type = "nn",
  pcs.use = 1:10,
  cells.use = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.use}{Dimenionality reduction used to calculate distances}

\item{dist.type}{How to calculate distances; 'nn' or 'centroid'}

\item{pcs.use}{PC dims used to calculate distances}

\item{cells.use}{subset of cells to calc distances; default NULL = all cells}
}
\description{
This function calculates next closest cluster for each cell; used for merging
}
\examples{
ComputeClusterDistances()
}
\keyword{binom}
