% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genes.R
\name{convert_genes}
\alias{convert_genes}
\title{Convert genes between symbol types}
\usage{
convert_genes(genes, from, to)
}
\arguments{
\item{genes}{The genes to convert.}

\item{from}{Gene symbol type of the input}

\item{to}{Desired gene symbol type for the output}
}
\value{
Data frame of genes with original and corresponding converted symbols
}
\description{
Takes an input of genes and converts from inputed type to desired type
}
