% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{de}
\alias{de}
\title{Differential expreesed analysis of two conditions.}
\usage{
de(mat, cond, qval = 0.1, contrast = NULL)
}
\arguments{
\item{mat}{Expression matrix.}

\item{cond}{Vector indicate the conditions of cells.}

\item{qval}{Significant level to select genes.}

\item{contrast}{How to compare.}
}
\value{
data.frame of DESeq2 results.
}
\description{
DE analysis based on DESeq2.
}
