% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{difAlignment}
\alias{difAlignment}
\title{Prop Expressing Test}
\usage{
difAlignment(object, gene, prefix, plot = FALSE, assay1, assay2)
}
\arguments{
\item{object}{Seurat object}

\item{gene}{Gene name}

\item{prefix}{Output file name prefix}

\item{group_by}{metadata variable to group cells by. defaults to Idents}

\item{split_by}{Further subset data by metadata variable}
}
\description{
Test to determine whether assays (alignment) have different proportions of GeneX+ cells per sample by cluster. Also tests for sex effect
}
\examples{
difAlignment()
}
\keyword{PercExp}
