% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkColData}
\alias{.checkColData}
\title{Checking colData names in SCE}
\usage{
.checkColData(sce, col)
}
\arguments{
\item{sce}{}

\item{col}{}
}
\value{
...
}
\description{
Checking colData names in SCE
}
