% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkEmbedding}
\alias{.checkEmbedding}
\title{Checking embedding SCE}
\usage{
.checkEmbedding(sce, embedding)
}
\arguments{
\item{sce}{}

\item{embedding}{}
}
\value{
...
}
\description{
Checking embedding SCE
}
