% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{euler_plot}
\alias{euler_plot}
\title{Euler plots}
\usage{
euler_plot(
  object,
  genes,
  metacol = NULL,
  metavar = NULL,
  title = NULL,
  assay = "RNA"
)
}
\arguments{
\item{object}{Seurat object}

\item{genes}{Gene names}

\item{metacol}{Column in metadata to subset cells by}

\item{metavar}{group in metadata column to subset cells by}

\item{title}{Title in plot}

\item{assay}{Seurat object assay}
}
\description{
This function generates euler plots from a seurat object using the eulerr package
}
\examples{
euler_plot()
}
\keyword{euler}
