% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{euler_plotv2}
\alias{euler_plotv2}
\title{Euler plot v2}
\usage{
euler_plotv2(object, genes, metacol = NULL, metavar = NULL, title = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{genes}{Gene names with assay info; e.g. c('rna_Snap25','exons_Snap25')}

\item{metacol}{Column in metadata to subset cells by}

\item{metavar}{group in metadata column to subset cells by}

\item{title}{Title in plot}
}
\description{
This function generates euler plots with eulerr package; can be used across multiple assays
}
\examples{
euler_plotv2()
}
\keyword{euler}
