% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_markers.R, R/filter_save_markers.R
\name{export_markers}
\alias{export_markers}
\alias{export_markers}
\title{Filter and save markers}
\usage{
export_markers(sce_object, k, padj = 0.01, auroc = 0.8, folder, filename)

export_markers(sce_object, k, padj = 0.01, auroc = 0.8, folder, filename)
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{k}{number of clusters}

\item{padj}{Adjusted p value}

\item{auroc}{Area Under Receiving Operating Curve}

\item{cluster}{number of cluster to select}

\item{vector_name}{Name to give the returned vector}

\item{sce_object}{A single cell experiment object}

\item{k}{number of clusters}

\item{cluster}{number of cluster to select}

\item{padj}{Adjusted p value}

\item{auroc}{Area Under Receiving Operating Curve}

\item{vector_name}{Name to give the returned vector}
}
\value{
Vector of marker genes

Vector of marker genes
}
\description{
Filters marker genes based on padj and k and saves them as a csv file

Filters marker genes based on padj and k and saves them as a csv file
}
