% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{extract_max}
\alias{extract_max}
\title{ExtractMax}
\usage{
extract_max(p)
}
\arguments{
\item{p}{plot}
}
\description{
This function is used in StackedVlnPlot to get max Y-axis values
}
\examples{
extract_max()
}
\keyword{extract}
