% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_PC_gene_weights.R
\name{extract_PC_gene_weights}
\alias{extract_PC_gene_weights}
\title{Extract principal component gene weights}
\usage{
extract_PC_gene_weights(sce_object, ntop = 500,
  file_name = "PCA_gene_weights.csv")
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{ntop}{Amount of genes to use for PCA}

\item{file_name}{Name for output csv file}
}
\value{
Extract PCA gene weights in dataframe pca_loadings and save as csv
}
\description{
Save heatmap in designated folder
}
