% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{FetchGO}
\alias{FetchGO}
\title{Fetch genes in GO term.}
\usage{
FetchGO(term, species = "mMus")
}
\arguments{
\item{term}{GO term id, forexample "GO:0001049".}

\item{species}{Mouse or human database to used. Mouse: mMus, Human: Hs}
}
\value{
Ensembl id of specifc GO term
}
\description{
Fetch ensembl ids of specific GO term id.
}
