% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{findDoublets}
\alias{findDoublets}
\title{Find doublets}
\usage{
findDoublets(object, cores = 1, pcs = 30, assay = "RNA")
}
\arguments{
\item{object}{Seurat object}

\item{cores}{Number of working cores for parallelization}

\item{pcs}{Number of PCs used in doubletfinder}

\item{assay}{Seurat assay to normalize/scale data on. value 'SCT' performs 'SCTransform' and increases processing time significantly}
}
\description{
This function uses the DoubletFinder package to identify and remove doublets; parallel computing supported
}
\examples{
findDoublets()
}
\keyword{doublets}
