% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_scoring.R
\name{gene_set_scoring}
\alias{gene_set_scoring}
\title{Scores cell expression of genes in set}
\usage{
gene_set_scoring(ser, geneset)
}
\arguments{
\item{ser}{A Seurat object to be scored by the gene set. Must contain scaled data}

\item{geneset}{A list of genes and their fold changes. Should be organized such that "genes" are genes and "FC" is (log)foldchange. All genes will be assumed to be significant.}
}
\value{
Outputs a vector of score values named as cell names
}
\description{
Takes in a gene set and uses normalized gene expression data to calculate a score describing extent to which
cells are expressing genes in the set.
}
\details{
Parameters:
}
