% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{getGOEnsembl}
\alias{getGOEnsembl}
\title{Retrieving genes in GO term.}
\usage{
getGOEnsembl(term, species = "mMus")
}
\arguments{
\item{term}{GO term id. For example, "GO:000830".}

\item{species}{Link to mouse(mMus) or human(Hs) database.}
}
\value{
A vector of Ensembl gene ids in given GO term.
}
\description{
Retrieving genes in GO term.
}
