% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{glm.enet}
\alias{glm.enet}
\title{Variable selction based on elasticnet penalty.}
\usage{
glm.enet(mat, cell_labels, alpha = 0.5, lam = "lambda.min",
  family = "multinomial", type.multinomial = "grouped", nfolds = 10,
  cv.plot = FALSE, ...)
}
\arguments{
\item{mat}{Gene expression matrix, columns are cells and rows are genes.}

\item{cell_labels}{cell clustering labels.}

\item{alpha}{The elastic-net mixing parameter.}

\item{lam}{Determine the lambda that according to cross-validation error - default is lambda.min. lambda.1se is also available.}

\item{family}{Response type. see \code{\link[glmnet]{glmnet}}.}

\item{type.multinomial}{If "grouped" then a grouped lasso penalty is used on the multinomial coefficients for a variable.}

\item{nfolds}{number of folds - default is 10.}

\item{cv.plot}{Save cross-validation plot.}

\item{...}{Additional arguments passed on to \code{\link[glmnet]{glmnet}}.}
}
\value{
Expression matrix with selected genes.
}
\description{
Fits a multinomial model via penalized maximum likelihood.
}
