% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{go_enrich}
\alias{go_enrich}
\title{Gene ontology enrichment analysis of gene clusters.}
\usage{
go_enrich(gene_labels, method = "weight01", save_dir = NA, nodeSize = 30,
  showNodes = 15, topNodes = 50, width = 12, height = 10)
}
\arguments{
\item{gene_labels}{Data frame of gene names with cluster labels.}

\item{method}{Test statistics to deal with the GO graph structure - default is weight01, a mixture between the elim and weight algorithms. Alternative algorithms are classic, elim, weight, lea, parentchild. See \code{\link[topGO]{topGO-package}} for more detail.}

\item{save_dir}{Where GO results to be saved. If NA results will be save in the folder called 'GO_results'.}

\item{nodeSize}{The nodes with less than nodeSize annotated genes are removed from the GO hierarchy}

\item{showNodes}{How many term to show in the bar plot.}

\item{width}{Width of the barplot.}

\item{height}{Height of the barplot.}
}
\value{
go_enrich returns a list of: 
\itemize{
  \item the statistic results of each clusters;
  \item the genes in each term of each cluster.
}
}
\description{
Perform GO analysis using topGO.
}
