% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCluster.R
\name{iforest}
\alias{iforest}
\title{Detect outliers.}
\usage{
iforest(mat, n_estimators = 1000, outliers_fraction = 0.1,
  random_state = 1, n_jobs = 1)
}
\arguments{
\item{mat}{Gene expression matrix, columns are cells and rows are genes.}

\item{n_estimators}{The number of base estimators in the ensemble.}

\item{outliers_fraction}{float in (0, 0.5). The amount of contamination of the data set, i.e. the proportion of outliers in the data set. Used when fitting to define the threshold on the decision function.}

\item{random_state}{RandomState instance.}

\item{n_jobs}{The number of jobs to run in parallel for both `fit` and `predict`. If -1, then the number of jobs is set to the number of cores.}

\item{verbose}{Controls the verbosity of the tree building process.}
}
\value{
list of outlier score and indice.
}
\description{
The implement of iforest is dependent on numpy, scipy and sklearn in python, make sure that they have been installed in your enviroment.
}
\details{
Use Isolation Forest algorithm to detect outliers of cells.
}
\references{
Liu, Fei Tony, Ting, Kai Ming and Zhou, Zhi-Hua. "Isolation forest." Data Mining, 2008. ICDM'08. Eighth IEEE International Conference on.

Liu, Fei Tony, Ting, Kai Ming and Zhou, Zhi-Hua. "Isolation-based anomaly detection." ACM Transactions on Knowledge Discovery from Data (TKDD) 6.1 (2012): 3.
}
