% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{IncMorphDif}
\alias{IncMorphDif}
\title{Incision Morphine analysis}
\usage{
IncMorphDif(
  object = seurat,
  module = "signature_1DEGIEG",
  prefix,
  gene_list,
  Sex = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{module}{module score metadata column name}

\item{prefix}{File name prefix}

\item{gene_list}{list of genes in module score}

\item{Sex}{Whether to test for sex as a variable}
}
\description{
Function for all acute pain vs morphine data analysis and plotting? too much likely
}
\examples{
IncMorphDif()
}
\keyword{Analysis}
\keyword{Morphine}
\keyword{Pain}
