% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdf.R
\name{is_tdf}
\alias{is_tdf}
\title{Test if the object is a}
\usage{
is_tdf(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `tdf` class.
}
\description{
This function returns `TRUE` for the object returned from the 
\code{generate.placebos} function.
and `FALSE` for all other objects, including regular data frames.
}
