% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdf.R
\name{is_tdf_multi}
\alias{is_tdf_multi}
\title{Test if the object is a tdf_multi object}
\usage{
is_tdf_multi(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `tdf_multi` class.
}
\description{
This function returns `TRUE` for the object returned from the 
\code{multiple.synth} function.
and `FALSE` for all other objects, including regular data frames.
}
