% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{label_transfer}
\alias{label_transfer}
\title{Transfers labels from reference data to query data}
\usage{
label_transfer(query, ref, refdata)
}
\arguments{
\item{query}{Seurat object to query for predictions}

\item{ref}{Seurat object to derive predictions from}

\item{refdata}{Reference labels to use for prediction (cell type, cluster identity, other metadata, etc)}
}
\value{
Outputs query Seurat object with predictions added as metadata
}
\description{
Generates predictions for query data from reference data labels by transferring labels from reference to query data
Parameters:
}
