% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{lpheatmap}
\alias{lpheatmap}
\title{Heatmap for gene expression pattern.}
\usage{
lpheatmap(mat, filename = NA, vmax = 2, vmin = -1, log = FALSE,
  standard = FALSE, hlight = NULL, cmap = NULL, cluster_rows = TRUE,
  cluster_cols = TRUE, gene_labels = NULL, cell_labels = NULL,
  gene_order = NULL, cell_order = NULL, gaps_row = NULL,
  gaps_col = NULL, plot_gaps_row = T, plot_gaps_col = T, ...)
}
\arguments{
\item{mat}{Expression matrix, columns are cells and rows are genes.}

\item{filename}{Filename to save your heatmap.}

\item{vmax}{The maximum value to plot.}

\item{vmin}{The minimum value to plot.}

\item{log}{log2 tranform the data.}

\item{standard}{Standardize the data, result in mean = 0 and std = 1 of each gene.}

\item{hlight}{The genes you want to highlight in the heatmap.}

\item{cmap}{The pallette of the heatmap, one can generate it using colorRampPalette().}

\item{cluster_rows}{boolean values determining if rows should be clustered or hclust object.}

\item{cluster_cols}{boolean values determining if columns should be clustered or hclust object.}

\item{gene_labels}{A vector of gene clustering labels.}

\item{cell_labels}{A vector of cell clustering labels.}

\item{gene_order}{Specify a sequence of gene clusters order, the values must be in gene_labels.}

\item{cell_order}{Specify a sequence of cell clusters order, the values must be in cell_labels.}

\item{gaps_row}{A vector indicates gaps of row, if NULL the gaps will be setting based on the gene_labels.}

\item{gaps_col}{A vector indicates gaps of column, if NULL the gaps will be setting based on the cell_labels.}

\item{plot_gaps_row}{Make gaps on row.}

\item{plot_gaps_col}{Make gaps on column.}

\item{...}{Additional arguments passed on to pheatmap.}
}
\value{
Ordering cluster results.
}
\description{
Plot the clustering result of cells and genes
}
