% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_de_plots.R
\name{make_de_plots}
\alias{make_de_plots}
\title{Runs differential expression and creates relevant plots for a ser object}
\usage{
make_de_plots(
  ser = NULL,
  list_ser = NULL,
  feats = NULL,
  out_dir = "2_de/",
  cols = NULL,
  feature_plots = TRUE,
  meta = NULL
)
}
\arguments{
\item{ser}{The Seurat object to use}

\item{list_ser}{List of DE seurat object}

\item{feats}{A subset of featurs to run DE (I.E. only surface markers)
If NULL then will run on all genes}

\item{out_dir}{Directory to write plots and save markers.}

\item{cols}{Name vector of colors by cluster}

\item{feature_plots}{Whether or not to include feature plots in the pdf output.}

\item{meta}{subset that you want to graph}
}
\value{
Creates plots in user inputted directory
}
\description{
Runs differential expression and creates relevant plots for a ser object
}
