% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gse_scores.R
\name{make_gse_scores}
\alias{make_gse_scores}
\title{Parameters:}
\usage{
make_gse_scores(
  ser,
  directory = NULL,
  out_dir = "3_gse",
  from_gene = "HGNC",
  to_gene = "MGI",
  subset_gse = NULL,
  csv_dir = NULL,
  type = "Real"
)
}
\arguments{
\item{ser}{Seurat object to process}

\item{directory}{Directory to get gmt file}

\item{out_dir}{Output directory}

\item{from_gene}{"ENSG" or "ENSMUSG" or "HGNC" or "MGI"}

\item{to_gene}{"MGI" or HGNC

This function will read in the genesets, either default or customized by users
and make feature plots from those genesets}

\item{subset_gse}{User defined subset of gene set (Optional)}

\item{csv_dir}{Directory to get csv file (Optional)}

\item{type}{Output score in Absolute or Real values}
}
\value{
Outputs seurat object
}
\description{
Parameters:
}
