% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_processing_plots.R
\name{make_processing_plots}
\alias{make_processing_plots}
\title{Make default plot array associated with processing step.
Parameters:}
\usage{
make_processing_plots(ser, out_dir = "1_process/", colors = NULL)
}
\arguments{
\item{ser}{Seurat object to process.}

\item{out_dir}{Output directory}

\item{colors}{Custom color scheme
  
Creates a number of plots and csv files relating to cluster composition
by all metadata in the ser object. This includes dim plots, feature plots,
and pie plots for cluster composition. Point size is automatically scaled
based on total number of cells and 10%/90% quantiles are used for all
feature plots. Reductions used are umap and phate.}
}
\description{
Make default plot array associated with processing step.
Parameters:
}
