% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tar_feat_plots.R
\name{make_tar_feat_plots}
\alias{make_tar_feat_plots}
\title{Make violin and feature plots for a given set of features}
\usage{
make_tar_feat_plots(
  ser,
  features,
  out_dir = "2_de",
  prefix = "tar_features",
  use_phate = TRUE
)
}
\arguments{
\item{ser}{Seurat file to use for plots}

\item{features}{Vector of features to make plots for}

\item{out_dir}{Directory to write plots}

\item{use_phate}{Boolean indicating whether to use phate dimensional reductions}
}
\description{
This function accepts a Seurat object and a vector of features to create
violin and feature plots. It chunks the vector into peices containing up to
9 features each. The output folder will contain a pdf of violin plots, each
page containing one chunk worth of features, and two png files for each
chunk - one feature plot of the chunk with no thresholding and one with
thresholding on the 10% and 90% quantile.
}
