% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tar_gene_plots.R
\name{make_tar_gene_plots}
\alias{make_tar_gene_plots}
\title{Make violin and feature plots for a given set of genes
Parameters:}
\usage{
make_tar_gene_plots(ser, genes, out_dir = "2_de", prefix = "2_de/tar_genes")
}
\arguments{
\item{ser}{Seurat file to use for plots}

\item{genes}{Vector of genes to make plots for}

\item{out_dir}{Directory to write plots}
}
\description{
Make violin and feature plots for a given set of genes
Parameters:
}
