% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{markers.binom.metagroup}
\alias{markers.binom.metagroup}
\title{Metagroup Binomial test}
\usage{
markers.binom.metagroup(
  object,
  metagroup,
  clust.1,
  clust.2 = NULL,
  effect.size,
  assay = "RNA",
  posFrac = 0.25
)
}
\arguments{
\item{object}{Seurat object}

\item{metagroup}{colname in object metadata}

\item{clust.1}{Group 1}

\item{clust.2}{Group 2; if NULL = all other cells}

\item{effect.size}{Filter DEGs by effect size}

\item{assay}{seurat object assay}
}
\description{
This function applies the markers binom function within specific metadata groups
}
\examples{
markers.binom.metagroup()
}
\keyword{metagroup}
