% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{markers.binom}
\alias{markers.binom}
\title{markers.binom}
\usage{
markers.binom(
  object,
  clust.1,
  clust.2 = NULL,
  effect.size,
  assay,
  posFrac = 0.25,
  posOnly = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{clust.1}{Group 1; a working group within seurat object levels/Idents}

\item{clust.2}{Group 2; if NULL = all other cells}

\item{effect.size}{Filter DEGs by effect size}

\item{assay}{Seurat object assay}

\item{posFrac}{Include only genes expressed by at least posFrac proportion of cells in either group; default = 25\%}
}
\description{
This function uses the binomial test function to identify differentially expressed genes
}
\examples{
markers.binom()
}
\keyword{markers}
