% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{merge.clusters.DE}
\alias{merge.clusters.DE}
\title{Merge clusters}
\usage{
\method{merge}{clusters.DE}(object, min.de.genes, effect.size, pval.cutoff, pcs.use, assay = "RNA")
}
\arguments{
\item{object}{Seurat object}

\item{min.de.genes}{Min # DEGs calculated by binom to keep clusters distinct}

\item{effect.size}{Min effect size to call DEG in binom test}

\item{pval.cutoff}{Max pvalue to call DEG in binom test}

\item{pcs.use}{If merging clusters, PCs used to reassign cells to nearest cluster}

\item{assay}{Seurat assay used}
}
\description{
This function merges clusters that don't have enough DEGs to delineate them
}
\examples{
merge.clusters.DE()
}
\keyword{binom}
