% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{modify_vlnplot}
\alias{modify_vlnplot}
\title{Modify Violin}
\usage{
modify_vlnplot(
  obj,
  feature,
  pt.size = 0,
  plot.margin = unit(c(-0.75, 0, -0.75, 0), "cm"),
  ...
)
}
\arguments{
\item{feature}{gene}

\item{pt.size}{point size}

\item{plot.margin}{Margins}

\item{object}{Seurat object}
}
\description{
This function is used in StackedVlnPlot for formatting
}
\examples{
modify_vlnplot()()
}
\keyword{violin}
