% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspe.plot.R
\name{mspe.plot}
\alias{mspe.plot}
\title{Plot the post/pre-treatment MSPE ratio}
\usage{
mspe.plot(tdf, discard.extreme = TRUE, mspe.limit = 20,
  plot.hist = FALSE, title = NULL, xlab = "Post/Pre MSPE ratio",
  ylab = NULL)
}
\arguments{
\item{tdf}{An object constructed by \code{\link{generate.placebos}}.}

\item{discard.extreme}{Logical. Whether or not placebos with high 
pre-treatement MSPE should be excluded from the plot.}

\item{mspe.limit}{Numerical. Used if \code{discard.extreme} is \code{TRUE}. 
It indicates how many times the pretreatment MSPE of a placebo should 
be higher than that of the treated unit to be considered extreme and 
discarded. Default is \code{20}.}

\item{plot.hist}{Logical. If \code{FALSE}, a dotplot with each unit name 
and its post/pre treatment MSPE ratio is produced. If \code{TRUE}, 
a histogram is produced, with the frequency of each ratio. 
Should be set to \code{TRUE} when there are many controls, to make 
visualization easier.}

\item{title}{Character. Optional. Title of the plot.}

\item{xlab}{Character. Optional. Label of the x axis.}

\item{ylab}{Character. Optional. Label of the y axis.}
}
\description{
Plots the post/pre-treatment mean square prediction error ratio 
    for the treated unit and placebos.
}
\details{
Post/pre-treatement mean square prediction error ratio is the 
    difference between the observed outcome of a unit and its synthetic 
    control, before and after treatement. A higher ratio means a small 
    pretreatment prediction error (a good synthetic control), and a high 
    post-treatment MSPE, meaning a large difference between the unit and 
    its synthetic control after the intervention. By calculating this ratio 
    for all placebos, the test can be interpreted as looking at how likely 
    the result obtained for a single treated case with a synthetic control 
    analysis could have occurred by chance given no treatement. 
    For more detailed description, see Abadie, Diamond, and Hainmueller (2011, 2014).
}
\references{
Abadie, A., Diamond, A., Hainmueller, J. (2014). 
  Comparative Politics and the Synthetic Control Method. 
  American Journal of Political Science Forthcoming 2014.

   Synthetic : An R Package for Synthetic Control Methods in Comparative 
   Case Studies. Journal of Statistical Software 42 (13) 1–17.

   Abadie, A., Diamond, A., Hainmueller, J. (2011). Synth: An R Package for 
   Synthetic Control Methods in Comparative Case Studies. 
   Journal of Statistical Software 42 (13) 1–17.

   Abadie A, Diamond A, Hainmueller J (2010). Synthetic Control Methods for 
   Comparative Case Studies: Estimating the Effect of California's Tobacco 
   Control Program. Journal of the American Statistical Association 
   105 (490) 493–505.
}
\seealso{
\code{\link{generate.placebos}}, \code{\link{mspe.test}}, 
    \code{\link{plot_placebos}}, \code{\link[Synth]{synth}}
}
