% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_scores.R
\name{organize_scores}
\alias{organize_scores}
\title{Sort matrix by Seurat metadata}
\usage{
organize_scores(scores, ser, meta)
}
\arguments{
\item{scores}{The matrix of values to be organized}

\item{ser}{The Seurat object containing the cells of interest}

\item{meta}{The metadata used to organize the values}
}
\value{
Organized matrix of scores
}
\description{
Organizes matrix of values associated with cells in a Seurat object 
by specified metadata
}
