% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{p.cor}
\alias{p.cor}
\title{Calculate the p-value of correlation coefficient.}
\usage{
p.cor(x, y = NULL, method = "pearson", correct = "BH", ...)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{y}{NULL (default) or a vector, matrix or data frame with compatible dimensions to x. The default is equivalent to y = x (but more efficient).}

\item{method}{A character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{correct}{Correction method for p-value, see \code{\link[stats]{p.adjust}}.}

\item{...}{Additional arguments passed on to \code{\link[stats]{cor}}.}
}
\value{
The matrix of the p-value of correlation coefficient.
}
\description{
Calculate the significant level of correlation coefficient.
}
