% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{PercExp}
\alias{PercExp}
\title{Percent Expressing}
\usage{
PercExp(object, features, assay = NULL, group_by = "NULL", split_by = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Gene names}

\item{assay}{Seurat object assay}

\item{group_by}{metadata variable to group cells by. defaults to Idents}

\item{split_by}{Further subset data by metadata variable}
}
\description{
This function calculates the percentage of cells expressing a given gene across groups and pads non-expressing groups with zeros.
}
\examples{
PercExp()
}
\keyword{PercExp}
