% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plac.dist.R
\name{plac.dist}
\alias{plac.dist}
\title{Plot the distribution of placebo samples for synthetic control 
    analysis with multiple treated units.}
\usage{
plac.dist(multiple.synth, nboots)
}
\arguments{
\item{multiple.synth}{multiple.synth}{An object returned by the function 
\code{\link{multiple.synth}}}

\item{nboots}{Number of bootstrapped samples of placebos to take.}
}
\value{
\describe{
\item{p}{The plot.}
\item{att.t}{The observed average treatment effect.}
\item{df}{Dataframe where each row is the ATT for one bootstrapped placebo 
   sample, used to build the distribution plot.}
}
}
\description{
Takes the output object of \code{\link{multiple.synth}} creates a 
    distribution of placebo average treatment effects, to test the 
    significance of the observed ATE. Does so by sampling k placebos 
    (where k = the number of treated units) nboots times, and calculating 
    the average treatment effect of the k placebos each time.
}
