% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plot_components}
\alias{plot_components}
\title{Plot components}
\usage{
plot_components(sce_object, PCx, PCy, group, folder, alpha = 0.8,
  palette = 2, gene = FALSE, width = 14, height = 10, units = "cm")
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{PCx}{Principal component to plot on x axis}

\item{PCy}{Principal component to plot on y axis}

\item{group}{Variable or gene name}

\item{folder}{Name of designated folder to save images}

\item{alpha}{Decide the transparency of geom}

\item{palette}{Number of palette selected from qualitative series on colorbrewer2.org}

\item{gene}{If TRUE it will plot genes and no variables}

\item{width}{Width of image in cm}

\item{height}{Height of image in cm}
}
\value{
PCA plot of single-cell dataset with variables or genes as color scale
}
\description{
Plot variables or genes of interest on a variable number of principal components
}
