% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dims.R
\name{plot_dims}
\alias{plot_dims}
\title{Plot reduced dimensions}
\usage{
plot_dims(sce_object, x, y, var, palette = "Dark2", hex_codes = NA,
  point_size = 3, alpha = 1, theme = 18)
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{x}{Principal component to plot on x axis}

\item{y}{Principal component to plot on y axis}

\item{var}{Variable or gene name}

\item{palette}{Number of palette selected from qualitative series on colorbrewer2.org}

\item{hex_codes}{Manuel entry of HEX color codes to plot}

\item{point_size}{Size parameter of geom_point}

\item{alpha}{Decide the transparency of geom}
}
\value{
Ggplot object plot of single-cell dataset with variables or genes as color scale
}
\description{
Plot reduced dimensions from singlecellexperiment object and var them on color or plot expression level
}
