% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot expression}
\usage{
plot_expression(sce_object, x, gene, group, folder, violin_fill = "#f0f0f0",
  brew_type = "qual", brew_palette = "Set1", point_size = 1.5,
  scatter_width = 0.2, theme = 12, width = 12, height = 10)
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{x}{Groups to plot on x axis}

\item{gene}{Principal component to plot on y axis}

\item{group}{Color variable}

\item{folder}{Name of designated folder to save image}

\item{violin_fill}{Fill color for violin}

\item{brew_type}{Kind color type to plot: sequential, diverging or qualitative}

\item{brew_palette}{Number of palette selected from qualitative series on colorbrewer2.org}

\item{point_size}{Size parameter of geom_point}

\item{scatter_width}{Measure how wide points should scatter}

\item{theme}{Number of theme_size}

\item{width}{Width of image in cm}

\item{height}{Height of image in cm}
}
\value{
Gene expression plot
}
\description{
Plot variables or genes of interest on a variable number of principal components
}
