% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_on_PCA.R
\name{plot_on_PCA}
\alias{plot_on_PCA}
\title{Plot on PCA}
\usage{
plot_on_PCA(sce_object, n, var)
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{n}{Amount of principal component to plot}

\item{var}{Vector of variable names to plot}
}
\value{
PCA plot of single-cell dataset with variables or genes as color scale
}
\description{
Plot variables or genes of interest on a variable number of principal components
}
