% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAnimatedEmbedding}
\alias{plotAnimatedEmbedding}
\title{plotAnimatedEmbedding}
\usage{
plotAnimatedEmbedding(
  sce,
  genes,
  dim = "UMAP",
  q = 0.95,
  assay.type = "logcounts",
  theme.args = NULL
)
}
\arguments{
\item{sce}{}

\item{genes}{}

\item{dim}{}

\item{q}{}

\item{theme.args}{}
}
\value{
plot
}
\description{
Plot gene expression in embedded space
}
