% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBarplots}
\alias{plotBarplots}
\title{plotBarplots}
\usage{
plotBarplots(
  sce,
  genes,
  color_by = "annotation",
  order_by = NULL,
  bins = NULL,
  assay.type = "logcounts"
)
}
\arguments{
\item{sce}{}

\item{genes}{}

\item{color_by}{}

\item{order_by}{}

\item{bins}{}

\item{assay.type}{}
}
\value{
plot
}
\description{
Quickly plot gene expression of genes as boxplots
}
