% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBoxplots}
\alias{plotBoxplots}
\title{plotGE}
\usage{
plotBoxplots(sce, genes, by)
}
\arguments{
\item{sce}{}

\item{genes}{}

\item{by}{}
}
\value{
plot
}
\description{
Quickly plot gene expression of genes as boxplots
}
