% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotEmbedding}
\alias{plotEmbedding}
\title{plotGeneExpression}
\usage{
plotEmbedding(
  sce,
  genes,
  dim = "UMAP",
  q = 0.95,
  average_expr = FALSE,
  only_average = FALSE
)
}
\arguments{
\item{sce}{}

\item{genes}{}

\item{dim}{}

\item{q}{}

\item{average_expr}{}

\item{only_average}{}
}
\value{
plot
}
\description{
Plot gene expression in embedded space
}
