% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_B.R
\name{process_B}
\alias{process_B}
\title{Processes_T: This function integrate B cell data with alpha and beta vdj sequence into the seurat object
Parameters:}
\usage{
process_B(ser, B_dir)
}
\arguments{
\item{ser}{Seurat object to process}

\item{B_dir}{Optional directory to B cell TCR}
}
\value{
Outputs a Seurat with T cell integration
}
\description{
Processes_T: This function integrate B cell data with alpha and beta vdj sequence into the seurat object
Parameters:
}
