% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_counts_hash.R
\name{process_counts_hash}
\alias{process_counts_hash}
\title{Processes counts into a Seurat object prepared for alignment. 
This file also seperated.Adds hashing information to a Seurat object and adds hash id to barcodes.
Parameters:}
\usage{
process_counts_hash(
  directory,
  from_gene,
  to_gene,
  hash_dir = NULL,
  hash_meta = NULL,
  sample_id = NULL,
  type = "DropSeq"
)
}
\arguments{
\item{directory}{directory containing the output files of align_dropseq_2.sh 
or cell ranger}

\item{from_gene}{"ENSG" or "ENSMUSG" or "HGNC" or "MGI"}

\item{to_gene}{"MGI" or "HGNC"}

\item{hash_dir}{Directory containing hash counts files}

\item{hash_meta}{csv file containing conversion of hash ID to sample name}

\item{sample_id}{Optional experimental ID to prefix cell names}

\item{type}{"DropSeq" or "10x"

This script reads in a counts file from either the DropSeq or 10x 
pipeline, converts the genes to a given naming convention (MGI or HGNC) and return a Seurat object}
}
\value{
Outputs a Seurat file prepared to rPCA anchoring
}
\description{
Processes counts into a Seurat object prepared for alignment. 
This file also seperated.Adds hashing information to a Seurat object and adds hash id to barcodes.
Parameters:
}
