% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_loom.R
\name{process_loom}
\alias{process_loom}
\title{Reads in loom files for rnavel_plot.R}
\usage{
process_loom(file, ser, hashed = FALSE)
}
\arguments{
\item{file}{Path to the output file of align_rnavel.sh. The name must match sample names in the provided metadata.csv file IF hashing was not performed. If hashing was performed then sample names will be determined from the hash meta data csv provided.}

\item{ser}{Seurat object containing target cells. If hashing is TRUE then it is recommended that ONLY the hashed cells are included in the given Seurat object. This is to prevent barcode overlap from different sequencing runs when searching for barcodes without the sample appended.}

\item{hashed}{Boolean indicating whether hashing was used to determine cell sample. If it was, then the hashing metadata file must be provided for the correct cell barcode format.}
}
\value{
Outputs a list of loom files from RNA velocity prepared for rnavel_plot.R or concatenation if processing multiple samples.
}
\description{
The function reads in loom files from align_rnavel.sh and formats the cell
names to match the naming convention of process_ser.R. The input files must
have the 'sample.loom' where sample is the correct sample name to append to
the cell barcode. Note that cells in the Seurat object must be named in the format
barcode_sample.
}
