% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_T.R
\name{process_T}
\alias{process_T}
\title{Processes_T: This function integrate T cell data with alpha and beta vdj sequence into the seurat object}
\usage{
process_T(ser, T_dir)
}
\arguments{
\item{ser}{Seurat object to process}

\item{T_dir}{Optional directory to T cell TCR}
}
\value{
Outputs a Seurat with T cell integration
}
\description{
Processes_T: This function integrate T cell data with alpha and beta vdj sequence into the seurat object
}
