% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plot.R
\name{pseudotime_plot}
\alias{pseudotime_plot}
\title{Creates pseudotime}
\usage{
pseudotime_plot(
  ser,
  out_dir = "4_pseudotime",
  cluster_start = NULL,
  cluster_end = NULL,
  cal_reduction = "phate",
  plot_reduction = "phate",
  group_by = NULL,
  lineage_ls = NULL
)
}
\arguments{
\item{ser}{Seurat object to process}

\item{out_dir}{Output directory}

\item{cluster_start}{Indicate what cluster to start with}

\item{cluster_end}{Indicate what cluster to end with}

\item{lineage_ls}{User customized list of cluster that want to be perform slingshot on}

\item{subset_gse}{User defined subset of gene set (Optional)}

\item{reduction}{Reduction method}
}
\description{
Creates pseudotime
}
