% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_seurat.R
\name{QualityFiltering}
\alias{QualityFiltering}
\title{Quality filtering}
\usage{
QualityFiltering(
  object,
  assay = "RNA",
  min_counts = 300,
  max_counts = 8000,
  min_feature = 200,
  max_feature = Inf,
  max_mito = 25
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Seurat assay to use for the quality filtering.}

\item{min_counts}{Minimum number of counts per, below which barcodes should be excluded.}

\item{max_counts}{Maximum number of counts, above which barcodes should be excluded.}

\item{min_feature}{Minimum number of features, below which barcodes should be excluded.}

\item{max_feature}{Maximum number of features, above which barcodes should be excluded.}

\item{max_mito}{Maximum percentage of mitochondrial genes, above which barcodes should be excluded.}
}
\value{

}
\description{
Quality filtering
}
