% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_seurat.R
\name{ReadMetaData}
\alias{ReadMetaData}
\title{Reads metadata from Excel file. Adds new columns but can also be used to
replace existing columns.}
\usage{
ReadMetaData(file, sheet)
}
\arguments{
\item{file}{Path of the excel file that contains the metadata.}

\item{sheet}{The excel sheet that should be read.}
}
\value{

}
\description{
Reads metadata from Excel file. Adds new columns but can also be used to
replace existing columns.
}
