% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnavel_plot.R
\name{rnavel_plot}
\alias{rnavel_plot}
\title{Creates an RNA velocity projection}
\usage{
rnavel_plot(
  loom,
  ser,
  out_file,
  dr = "phate",
  cols = NULL,
  n_core = NULL,
  plot_info = NULL,
  ...
)
}
\arguments{
\item{loom}{List of loom files as read in from process_loom.R.}

\item{ser}{Seurat object containing clusters and dimensional reduction.}

\item{out_file}{File name of png file to save plot.}

\item{dr}{Name of dimensional reduction slot to use from Seurat object. Defaults to phate.}

\item{cols}{Named vector of colors for cell clusters. If NULL then ggplot colors are generated.}

\item{n_core}{Number of cores to use for velocyto.R functions. If NULL then future::availableCores will be used.}

\item{plot_info}{If RNA velocity was run previously you can provide the output file to skip much of the processing time. This can be very helpful when tweaking plotting parameters.}

\item{...}{Parameters to pass to velocyto.R::show.velocity.on.embedding.cor}
}
\value{
Return lists of objects containing information about rna velocity
}
\description{
The function takes a loom file and Seurat object and creates a projection of
RNA velocity onto a dimensional reduction of your choice. The dimensional
reduction should be present in the Seurat object.
}
