% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_PC_gene_weights.R
\name{save_PC_gene_weights}
\alias{save_PC_gene_weights}
\title{Save Principal component weights}
\usage{
save_PC_gene_weights(df, PC)
}
\arguments{
\item{df}{Dataframe with genes as rownames and PC weights as columns}

\item{PC}{Principal component}
}
\value{
Save csv files of negative and positive gene weights per principal component
}
\description{
Save heatmap in designated folder
}
