% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pheatmap_tiff.R
\name{save_pheatmap_tiff}
\alias{save_pheatmap_tiff}
\title{Save heatmap}
\usage{
save_pheatmap_tiff(x, filename, width = 10, height = 10)
}
\arguments{
\item{x}{Object}

\item{filename}{Name of output file}

\item{width}{Width of image in cm}

\item{height}{Height of image in cm}
}
\value{
PCA plot of single-cell dataset with variables or genes as color scale
}
\description{
Save heatmap in designated folder
}
