% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClusterCompare.R
\name{scClusterCompare}
\alias{scClusterCompare}
\title{Comparing FindMarkers results}
\usage{
scClusterCompare(known_markers, novel_markers)
}
\arguments{
\item{x}{\itemize{
\item the validated markers
}}

\item{y}{\itemize{
\item your newly generated markers from a new dataset
}}
}
\description{
This function allows you to compare the results of the FindMarkers function from the Seurat package.
This function is for people who have one set of markers that they feel are validated for a specific cell type
and want to reuse them for annotation in a new dataset. If you do not know, which marker genes were crucial for the annotation
this can be difficult without the help of a biologist. That is why this package allows to automize this process, by calculating
a score that can to some degree show which clusters are similar to which.
}
\details{
The function returns a dataframe with n number of columns, corresponding to the number of original dataset clusters plus one with the score.
The number of rows corresponds to the number of clusters in the new dataset.
}
\keyword{Seurat,}
\keyword{cell}
\keyword{markers}
\keyword{type,}
