% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCorrelate.R
\name{scCorrelate}
\alias{scCorrelate}
\title{Creating a dataframe of correlated genes to our genes of interest}
\usage{
scCorrelate(Seurat, gene_symbols)
}
\arguments{
\item{Seurat}{\itemize{
\item Seurat object
}}

\item{gene_symbols}{\itemize{
\item genes of interest
}}

\item{assay}{\itemize{
\item choose Seurat assay
}}

\item{p_val}{\itemize{
\item p-value cutoff for FDR
}}
}
\description{
Creating a dataframe of correlated genes to our genes of interest
}
\keyword{Seurat,}
\keyword{cell,}
\keyword{correlation}
\keyword{single}
