% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCorrelateLite.R
\name{scCorrelateLite}
\alias{scCorrelateLite}
\title{Creating a dataframe of correlated genes to our genes of interest}
\usage{
scCorrelateLite(Seurat, gene_symbols)
}
\arguments{
\item{Seurat}{\itemize{
\item Seurat object
}}

\item{genes}{\itemize{
\item genes of interest
}}

\item{assay}{\itemize{
\item choose Seurat assay
}}

\item{corr_test}{\itemize{
\item which type of correlation test
}}

\item{p_val}{\itemize{
\item p-value cutoff for FDR
}}
}
\description{
Creating a dataframe of correlated genes to our genes of interest
}
\keyword{Seurat,}
\keyword{cell,}
\keyword{correlation}
\keyword{single}
