% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCountExtract.R
\name{scCountExtract}
\alias{scCountExtract}
\title{Extracting counts from the Seurat object to use other types of analysis}
\usage{
scCountExtract(Seurat, use_metadata, features)
}
\arguments{
\item{Seurat}{\itemize{
\item Seurat object
}}

\item{use_metadata}{\itemize{
\item TRUE/FALSE whether you want to extract specific metadata
}}

\item{features}{\itemize{
\item metadata you want to use
}}
}
\description{
Extracting counts from the Seurat object to use other types of analysis
}
\keyword{Seurat,}
\keyword{cell,}
\keyword{correlation}
\keyword{single}
