% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCreateRawSamples.R
\name{scCreateRawSamples}
\alias{scCreateRawSamples}
\title{Creating a list of scRNA-seq objects for further analysis}
\usage{
scCreateRawSamples(x = getwd(), min.cells, min.features)
}
\arguments{
\item{x}{\itemize{
\item The directory of your files
}}
}
\description{
This function allows you to combine raw scRNA-seq files into a list of Seurat objects.
}
\keyword{Seurat}
\keyword{objecy}
