% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDoublets.R
\name{scDoublets}
\alias{scDoublets}
\title{Finding doublets in Seurat Objects}
\usage{
scDoublets(Seurat_Object)
}
\arguments{
\item{Seurat_Object}{\itemize{
\item Seurat object
}}
}
\description{
Finding doublets in Seurat Objects
}
