% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFeatureSelection.R
\name{scFeatureSelection}
\alias{scFeatureSelection}
\title{Feature selection using the Monte Carlo Feature Selection algorithm}
\usage{
scFeatureSelection(Seurat, use_metadata, features)
}
\description{
Feature selection using the Monte Carlo Feature Selection algorithm
}
\keyword{Seurat,}
\keyword{cell,}
\keyword{correlation}
\keyword{single}
