% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFeatureSelectionLite.R
\name{scFeatureSelectionLite}
\alias{scFeatureSelectionLite}
\title{Feature selection using gradient boosting}
\usage{
scFeatureSelectionLite(Seurat, use_metadata, features)
}
\description{
Feature selection using gradient boosting
}
\keyword{Seurat,}
\keyword{cell,}
\keyword{correlation}
\keyword{single}
