% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFilter.R
\name{scFilter}
\alias{scFilter}
\title{Use the results of IntroQC for quality control application}
\usage{
scFilter(QC_results, Seurat_Object)
}
\arguments{
\item{QC_results}{\itemize{
\item results of IntroQC
}}

\item{Seurat}{\itemize{
\item Seurat object
}}
}
\description{
Use the results of IntroQC for quality control application
}
\keyword{cell,}
\keyword{control}
\keyword{quality}
\keyword{single}
