% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegrate.R
\name{scIntegrate}
\alias{scIntegrate}
\title{Integrating scRNA-seq samples into a single Seurat Object}
\usage{
scIntegrate(Seurat_Object)
}
\arguments{
\item{Seurat}{\itemize{
\item Seurat object
}}
}
\description{
Integrating scRNA-seq samples into a single Seurat Object
}
