% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntroQC.R
\name{scIntroQC}
\alias{scIntroQC}
\title{scRNA-seq quality control checks}
\usage{
scIntroQC(x)
}
\arguments{
\item{x}{\itemize{
\item the Seurat file (or list of files)
}}
}
\description{
This function calculates a couple of basic quality control metrics and returns a plot containing
multiple violin plots with the mean and median of:
\itemize{
\item nFeature_RNA
\item percentage of mitochondrial genes in cell
\item percentage of ribosomal genes in cell
\item percentage of red blood cell genes
}
}
\details{
The function returns the Seurat file (or list of files) with added quality control metrics and a
plot containing the quality control metrics as violin plots.
To select the best object in a pre-integrated list of cells, the file with the highest number of cells will be
used as a benchmark.
}
\keyword{Seurat,}
\keyword{control}
\keyword{quality}
