% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_prediction.R
\name{scn_prediction}
\alias{scn_prediction}
\title{singleCellNet training and prediction}
\usage{
scn_prediction(query, stRef, expRef)
}
\arguments{
\item{query}{The data which will be classified by the random forest classifier}

\item{stRef}{Reference sample data. Needs to have a column named "cell" with cell names and one 
named "newAnn" with annotations/desired clasifications.}

\item{expRef}{Reference experiment data. Should contain genes as row names, cell names as column
names and contain gene expression levels in the matrix}
}
\value{
Returns a list with the trained classifier, predictions for the queried data, 
                 and assessment information from classifier training.
}
\description{
Trains a singleCellNet classifier and then predicts classification of query data
}
