% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtools.R
\docType{package}
\name{SCtools-package}
\alias{SCtools}
\alias{SCtools-package}
\title{SCTools: Tools for Synthetic Control Methods}
\description{
A set of functions to extend the synthetic controls analyses 
performed by the package Synth. Includes generating and plotting placebos, 
significance tests and plots, and calculating average treatment effects 
for multiple treated units.
}
\details{
It has several goals:

\itemize{
\item Allow easy generation of placebos
\item Generate figures for inference on SCM outputs
\item Extend the existing Synth package
}
}
\author{
\strong{Maintainer}: Bruno Castanho Silva \email{bcsilva@wiso.uni-koeln.de} (YOUR-ORCID-ID)

Authors:
\itemize{
  \item Michael DeWitt \email{me.dewitt.jr@gmail.com} (0000-0001-8940-1967)
}

}
