% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVisualize.R
\name{scVisualize}
\alias{scVisualize}
\title{Creating basic Seurat visualizations as wel as a heatmap if your choice is so}
\usage{
scVisualize(Seurat_Object, features, create_heatmap, heatmap_grouping)
}
\arguments{
\item{features}{\itemize{
\item genes of interest
}}

\item{create_heatmap}{\itemize{
\item choose Seurat assay
}}

\item{heatmap_grouping}{\itemize{
\item metadata
}}

\item{Seurat}{\itemize{
\item Seurat object
}}
}
\description{
Creating basic Seurat visualizations as wel as a heatmap if your choice is so
}
\keyword{Seurat,}
\keyword{cell,}
\keyword{correlation}
\keyword{single}
