% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_markers.R
\name{select_markers}
\alias{select_markers}
\title{Select markers}
\usage{
select_markers(sce_object, k, cluster, padj = 0.01, auroc = 0.8,
  vector_name = "genes")
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{k}{number of clusters}

\item{cluster}{number of cluster to select}

\item{padj}{Adjusted p value}

\item{auroc}{Area Under Receiving Operating Curve}

\item{vector_name}{Name to give the returned vector}
}
\value{
Vector of marker genes
}
\description{
Select marker genes based on padj, auroc, cluster number and k and returns them as a vector
}
