% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratHelpR.R
\name{sexDif}
\alias{sexDif}
\title{Sex differences}
\usage{
sexDif(object, gene, prefix, assay, plot = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{gene}{Gene name}

\item{prefix}{File name prefix}

\item{assay}{seurat object assay}

\item{plot}{Whether to generate box plots split by sex}
}
\description{
test whether there are differences in the proportion of GeneX+ cells between sexes by sample; generates plots of significant differences (p val <.05 or abs(effect size) > 1) by cluster. metadata column 'Sex' with values 'Male' or 'Female' must exist.
}
\examples{
sexDif()
}
\keyword{Sex}
\keyword{differences}
