% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCluster.R
\name{sig.pcs}
\alias{sig.pcs}
\title{Discover significant latent variation in data.}
\usage{
sig.pcs(mat, B = 20, threshold = 0.05, log = TRUE, scale = TRUE)
}
\arguments{
\item{mat}{Gene expression matrix, columns are cells and rows are genes.}

\item{B}{A number (a positive integer) of resampling iterations.}

\item{threshold}{A numeric value between 0 and 1 to threshold p-values.}

\item{log}{Comppute PCA of logarithm.}

\item{scale}{Scale the gene expression.}
}
\value{
Number of significant pcs by permutation parallen analysis.
}
\description{
Calculate statistical significance of latent variation using jackstraw.
}
