% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCluster.R
\name{sig.var}
\alias{sig.var}
\title{Discover high discriminant genes in unsupervised manner}
\usage{
sig.var(mat, r = NULL, s = 100, B = 400, adj.method = "fdr",
  log = TRUE, scale = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{mat}{Gene expression matrix, columns are cells and rows are genes.}

\item{r}{A number (a positive integer) of significant principal components.}

\item{s}{A number (a positive integer) of “synthetic” null variables. Out of m variables, s variables are independently permuted.}

\item{B}{A number (a positive integer) of resampling iterations. There will be a total of s*B null statistics.}

\item{adj.method}{Correction method for p-values, see \code{\link[stats]{p.adjust}}.}

\item{log}{Comppute PCA of logarithm.}

\item{scale}{Scale the gene expression.}

\item{verbose}{A logical specifying to print the computational progress.}

\item{...}{Additional arguments passed on to \code{\link[jackstraw]{jackstraw_pca}}.}
}
\value{
Statistical results of jackstraw.
}
\description{
Calculate statistical significance of variables driving latent variation using jackstraw
}
