% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_set_scoring.R
\name{signed_set_scoring}
\alias{signed_set_scoring}
\title{Scores cell expression of genes in set taking into account sign of fold change}
\usage{
signed_set_scoring(
  ser,
  geneset,
  from_gene = "MGI",
  to_gene = "MGI",
  scaled = TRUE
)
}
\arguments{
\item{ser}{A Seurat object to be scored by the gene set. Must contain scaled data}

\item{geneset}{A list of genes and their fold changes.}

\item{from_gene}{Original gene type}

\item{to_gene}{Gene type for conversion}

\item{scaled}{Boolean to determine whether to scale score by the number of genes in the set}
}
\value{
Outputs a vector of score values named as cell names
}
\description{
Takes in a gene set and uses normalized scaled gene expression data to calculate a score describing extent to which
cells are expressing genes in the set. Seurat object must contain scale_data assay. Gene set must be organized
such that "genes" are genes and "FC" is (log)foldchange. All genes will be assumed to be significant.
}
