% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCluster.R
\name{snn_cliq}
\alias{snn_cliq}
\title{SNN-Cliq clustering.}
\usage{
snn_cliq(dmat, knn = 3, merge_cutoff = 0.5, r_cutoff = 0.7)
}
\arguments{
\item{dmat}{Paire-wised distance of cells.}

\item{knn}{Number of nearest neighbors to calculate shared nearest neighbors as secondary distance.}

\item{merge_cutoff}{merge_cutoff threshold.}

\item{r_cutoff}{r_cutoff threshold.}
}
\value{
Clucstering labels
}
\description{
Clustering cells using SNN-Cliq algorithm.
}
