% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tclone_processing.R
\name{Tclone_processing}
\alias{Tclone_processing}
\title{This file process T cell clonality.}
\usage{
Tclone_processing(ser, out_dir = "Cell/")
}
\arguments{
\item{ser}{Seurat object containing T cell information}

\item{out_dir}{Output directory 

This script reads in a counts file from either the DropSeq or 10x 
pipeline, converts the genes to a given naming convention (MGI or HGNC) and return a Seurat object}
}
\description{
Parameters:
}
