% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePattern.R
\name{violinPlot}
\alias{violinPlot}
\title{Violin plot.}
\usage{
violinPlot(mat, gene, level = NULL, levels = NULL, color = NULL,
  log = TRUE, jitter = FALSE)
}
\arguments{
\item{mat}{Expression matrix.}

\item{gene}{Gene id in rownames of mat.}

\item{level}{Level in colnames of mat.}

\item{levels}{Factor, custom level define by user.}

\item{color}{Vector of rgb color indices for violin.}

\item{log}{Boolen value, log transform the expression or not.}

\item{jitter}{Boolen value, plot dots or not.}
}
\value{
ggplot2 object
}
\description{
Violin plot for specific gene in expression matrix.
}
