% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Alcohol Data}
\format{a data.frame with 5107 rows and 8 columns:
\describe{
  \item{country_name}{The name of the country}
  \item{year}{year}
  \item{consumption}{name of the experimental unit}
  \item{country_code}{outcome of interest}
  \item{labor_force_participation_rate}{Covariate 1}
  \item{mobile_cellular_subscriptions}{Covariate 2}
  \item{inflation}{Covariate 3}
  \item{manufacturing}{Covariate 3}
  \item{country_num}{Covariate 3}
   ...
}}
\source{
World Health Organization, World Bank
}
\usage{
alcohol
}
\description{
Alcohol Data
}
\keyword{datasets}
