% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_sers.R
\name{align_sers}
\alias{align_sers}
\title{Aligns a list of ser objects}
\usage{
align_sers(sers, meta_file = "metadata.csv", ref = NULL)
}
\arguments{
\item{sers}{A list of Seurat objects to align. The objects should be normalized with variable genes ID'd.}

\item{meta_file}{Path to a metadata file containing all metadata by sample. Cell barcodes should be sample_name-XXXXXXXX and the metadata file should have sample_name as the first column.}

\item{ref}{NULL or integer specifying which ser in the sers list to use as a reference. If NULL then all combinations will be calculated and the best integration pass selected.}
}
\value{
Integrated Seurat object with mnn components as pca reduction.
}
\description{
This function integrates separate batches of cells typically processed by
process_counts_hash.
}
