% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_markers.R
\name{export_markers}
\alias{export_markers}
\title{Filter and save markers}
\usage{
export_markers(sce_object, k, padj = 0.01, auroc = 0.85, folder,
  filename)
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{k}{number of clusters}

\item{padj}{Adjusted p value}

\item{auroc}{Area Under Receiving Operating Curve}

\item{folder}{Folder where to export the file to}

\item{filename}{Name for exported csv file}
}
\value{
csv file
}
\description{
Filters marker genes based on padj and k and saves them as a csv file
}
