% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plot_components}
\alias{plot_components}
\title{Plot components}
\usage{
plot_components(sce_object, PCx, PCy, group, gene = FALSE, save = TRUE,
  folder, brewer = TRUE, palette = 2, hex_codes, point_size = 3,
  alpha = 0.8, theme = 18, width = 14, height = 10, units = "cm")
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{PCx}{Principal component to plot on x axis}

\item{PCy}{Principal component to plot on y axis}

\item{group}{Variable or gene name}

\item{gene}{If TRUE it will plot genes and no variables}

\item{folder}{Name of designated folder to save image}

\item{brewer}{If TRUE it will use colorpalettes from colorbrewer2.org}

\item{palette}{Number of palette selected from qualitative series on colorbrewer2.org}

\item{hex_codes}{Manuel entry of HEX color codes to plot}

\item{point_size}{Size parameter of geom_point}

\item{alpha}{Decide the transparency of geom}

\item{width}{Width of image in cm}

\item{height}{Height of image in cm}

\item{units}{Units to plot image in}
}
\value{
PCA plot of single-cell dataset with variables or genes as color scale
}
\description{
Plot variables or genes of interest on a variable number of principal components
}
