% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dims.R
\name{plot_dims}
\alias{plot_dims}
\title{Plot reduced dimensions}
\usage{
plot_dims(sce_object, x = "PC1", y = "PC2", color, shape = NA,
  labels = NA, col_values = NA, point_size = 3, alpha = 1, theme = 12,
  label_size = 18, nrow = NULL, ncol = NULL)
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{x}{Principal component to plot on x axis}

\item{y}{Principal component to plot on y axis}

\item{color}{factor for color}

\item{shape}{factor for shape}

\item{labels}{list containing labels for gene names}

\item{col_values}{Manuel entry of colors or HEX color codes to plot}

\item{point_size}{Size parameter of geom_point}

\item{alpha}{Decide the transparency of geom}

\item{theme}{numerical specifying base size of theme}

\item{label_size}{numerical specifying size of facet titles ie genes}

\item{nrow}{numerical specifying the amount of rows}

\item{ncol}{numerical specifying the amount of columns}
}
\value{
Ggplot object plot of single-cell dataset with variables or genes as color scale
}
\description{
Plot reduced dimensions from singlecellexperiment object and var them on color or plot expression level
}
