% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot expression}
\usage{
plot_expression(sce_object, var, group, theme = 12, facet = "vertical",
  type = "violin", scale = "width")
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{var}{Gene or genes to plot}

\item{group}{Grouping variable}

\item{theme}{Size of theme}

\item{facet}{Orientation of facets}

\item{type}{type of plot, option for violin or boxplot}

\item{scale}{Set scale of violin, if "area" (default), all violins have the same area (before trimming the tails). If "count", areas are scaled proportionally to the number of observations. If "width", all violins have the same maximum width.}
}
\value{
Gene expression plot
}
\description{
Plot variables or genes of interest on a variable number of principal components
}
