% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot expression}
\usage{
plot_expression(sce_object, var, group, theme = 12)
}
\arguments{
\item{sce_object}{A single cell experiment object}

\item{var}{Gene or genes to plot}

\item{group}{Grouping variable}

\item{theme}{Size of theme}
}
\value{
Gene expression plot
}
\description{
Plot variables or genes of interest on a variable number of principal components
}
