% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBoth}
\alias{plotBoth}
\title{plotBoth}
\usage{
plotBoth(sce, genes, by, dim = "UMAP", q = 0.95)
}
\arguments{
\item{sce}{}

\item{genes}{}

\item{by}{}

\item{dim}{}

\item{q}{}
}
\value{
plot
}
\description{
Plot gene expression in embedded space and a boxplot below
}
