% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotEmbedding}
\alias{plotEmbedding}
\title{plotEmbedding}
\usage{
plotEmbedding(
  sce,
  by,
  dim = "UMAP",
  q = 0.95,
  average_expr = FALSE,
  only_average = FALSE,
  assay.type = "logcounts",
  theme.args = NULL,
  return_plotlist = FALSE
)
}
\arguments{
\item{sce}{}

\item{by}{}

\item{dim}{}

\item{q}{}

\item{average_expr}{}

\item{only_average}{}

\item{assay.type}{}

\item{theme.args}{}

\item{return_plotlist}{}
}
\value{
plot
}
\description{
Plot gene expression in embedded space
}
