% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_B.R
\name{process_B}
\alias{process_B}
\title{This function integrates B cell data vdj sequence into the seurat object}
\usage{
process_B(ser, B_dir)
}
\arguments{
\item{ser}{Seurat object to process}

\item{B_dir}{Optional directory to B cell BCR}
}
\value{
Outputs a Seurat with B cell integration
}
\description{
This function integrates B cell data vdj sequence into the seurat object
}
