% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_counts_hash.R
\name{process_counts_hash}
\alias{process_counts_hash}
\title{Processes counts into a Seurat object prepared for alignment.}
\usage{
process_counts_hash(
  directory,
  from_gene,
  to_gene,
  hash_dir = NULL,
  hash_meta = NULL,
  sample_id = NULL,
  type = "DropSeq",
  T_dir = NULL,
  B_dir = NULL,
  umi_thresh = 500
)
}
\arguments{
\item{directory}{directory containing the output files of align_dropseq_2.sh 
or cell ranger}

\item{from_gene}{"ENSG" or "ENSMUSG" or "HGNC" or "MGI"}

\item{to_gene}{"MGI" or "HGNC"}

\item{hash_dir}{Directory containing hash counts files}

\item{hash_meta}{csv file containing conversion of hash ID to sample name}

\item{sample_id}{Optional experimental ID to prefix cell names}

\item{type}{"DropSeq" or "10x"}

\item{T_dir}{Optional directory to T cell TCR}

\item{B_dir}{optional directory to B cell BCR}

\item{umi_thresh}{Minimum total umi count for cells}
}
\value{
Outputs a Seurat file prepared to rPCA anchoring
}
\description{
This script reads in a counts file from either the DropSeq or 10x 
pipeline, converts the genes to a given naming convention (MGI or HGNC) and creates Seurat object from counts.
It is the input for align_sers. This file also adds hashing information to a Seurat object and adds hash id to barcodes.
}
