% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plot3d.R
\name{pseudotime_plot3d}
\alias{pseudotime_plot3d}
\title{Generate flashing and spinning 3d pseudotime plots}
\usage{
pseudotime_plot3d(
  ser,
  curve_dir,
  pseudotime_dir,
  subset = NULL,
  out_dir = "5_pseudotime_3d",
  bg_color = "white",
  frame = 500
)
}
\arguments{
\item{ser}{Seurat object to process}

\item{curve_dir}{Input curve directory}

\item{pseudotime_dir}{Input pseudotime directory}

\item{subset}{A subset of clusters that user want to see}

\item{out_dir}{Output directory}

\item{bg_color}{Background color}

\item{frame}{Number of frames}
}
\value{
Makes png files that can be assembled into 3d moving plots
}
\description{
Generate flashing and spinning 3d pseudotime plots
}
