% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_de.R
\name{run_de}
\alias{run_de}
\title{Runs differential expression and creates relevant plots for a ser object
Parameters:}
\usage{
run_de(ser, feats = NULL, out_dir = "2_de/", meta = NULL, origin = "Condition")
}
\arguments{
\item{ser}{The Seurat object to use}

\item{feats}{A subset of featurs to run DE (I.E. only surface markers)
If NULL then will run on all genes}

\item{out_dir}{Directory to write plots and save markers.}

\item{meta}{(Optional) Pass in cluster name that user wants to run DE}

\item{origin}{(Optional) Pass in what parameters in the each clusters that user wants to run DE}

\item{res}{(Optional) Resolution for clustering in each subset}
}
\value{
Output a processed differenttial expression for plotting
}
\description{
Runs differential expression and creates relevant plots for a ser object
Parameters:
}
