% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_de.R
\name{run_de}
\alias{run_de}
\title{Runs differential expression for a ser object}
\usage{
run_de(
  ser,
  feats = NULL,
  out_dir = "2_de/",
  meta = NULL,
  origin = "Condition",
  verbose = F
)
}
\arguments{
\item{ser}{The Seurat object to use}

\item{feats}{A subset of featurs to run DE (I.E. only surface markers)
If NULL then will run on all genes}

\item{out_dir}{Directory to write plots and save markers.}

\item{meta}{(Optional) Pass in cluster name that user wants to run DE}

\item{origin}{(Optional) Pass in what parameters in the each clusters that user wants to run DE}

\item{verbose}{Boolean flag for verbose output}
}
\value{
Output a processed differential expression for plotting
}
\description{
Creates object used in make_de_plots function for the creation of various plots.
Can be used on whole data set or individual clusters.
}
