\name{SCtools-package}
\alias{SCtools-package}
\alias{SCtools}
\docType{package}
\title{
\packageTitle{SCtools}
}
\description{
\packageDescription{SCtools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SCtools}
\packageIndices{SCtools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SCtools}

Maintainer: \packageMaintainer{SCtools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
