% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tclone_processing.R
\name{Tclone_processing}
\alias{Tclone_processing}
\title{This file process T cell clonality.}
\usage{
Tclone_processing(ser)
}
\arguments{
\item{ser}{Seurat object containing T cell information}
}
\value{
Feature plots highlighting clones
}
\description{
This script reads in a counts file from either the DropSeq or 10x 
pipeline, converts the genes to a given naming convention (MGI or HGNC) and return a Seurat object
}
